Finalize refactor: replace app and HTML with new version
