+++ version.json
@@ -1,4 +1,4 @@
 {
-  "commitHash": "f928350",
-  "buildTime": "2025-04-16T01:22:00.109Z"
+  "commitHash": "32a85a6",
+  "buildTime": "2025-04-16T01:43:28.792Z"
 }
